[Refactor][Quantization] Reorganize quantization module
